LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-remove-head-expected.txt


FAIL Meta color-scheme applies. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "auto"
FAIL Initial value after removing head including meta color-scheme. assert_equals: expected "light" but got "dark"
